requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
redis==4.3.4
openai==0.27.0
python-dotenv==0.19.0
pydantic==1.9.0
discord.py==2.0.1
boto3==1.26.137
celery==5.2.7
mangum==0.17.0
pydantic-settings==2.0.1
pytest==7.3.1
pytest-asyncio==0.21.0
httpx==0.24.1
moto==4.1.11
pytest-cov==4.1.0
requests==2.31.0
pytest-celery==0.0.0
pytest-mock==3.10.0
PyNaCl==1.5.0 